Split examples/demo.py into multiple modules
- this makes it easier to read the examples in order of complexity
- also useful later on when checking against the schema validator
